+++ src/repo_analyzer/render.clj
@@ -42,6 +42,7 @@
                              :href        "https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css"
                              :integrity   "sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T"
                              :crossorigin "anonymous"}]
+                     [:script {:src "https://cdnjs.cloudflare.com/ajax/libs/Chart.js/2.8.0/Chart.bundle.min.js"}]
                      ]
                     ; TODO include JS if needed for bootstrap
                     ;<script src="https://code.jquery.com/jquery-3.3.1.slim.min.js" integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo" crossorigin="anonymous"></script>
@@ -119,12 +120,22 @@
         all-commits-filename (string/join [base-path "all-commits-list.html"])
         self-commits-filename (string/join [base-path "self-committed-list.html"])
         different-committer-filename (string/join [base-path "different-committer-list.html"])
+        pie-chart-dataset (string/join "," [
+                                            (get-in analysis [:commit-statistics :self-committed :count])
+                                            (get-in analysis [:commit-statistics :committed-by-different-dev :count])])
         ]
     (create-site all-commits-filename "All commits" commit-list-html)
     (create-site self-commits-filename "Self committed commits" self-commit-list-html)
     (create-site different-committer-filename "Commits where committer and author are different" different-committer-list-html)
     (html
       [:h1 "Commit analysis"]
+      [:canvas {:id "commit-chart" :width "770px" :height "385px"}]
+      [:script "new Chart('commit-chart',
+      {'type':'pie','data':{'labels':['Self committed','Committer and author are different'],
+      'datasets':[{'label':'Committed / Authored','data': [" pie-chart-dataset "],
+      'backgroundColor':['rgb(255, 99, 132)','rgb(54, 162, 235)','rgb(255, 205, 86)']}]},
+      'options': {'responsive': false}});
+      "]
       [:p "Commits analyzed: " (:count (:commit-statistics analysis))
        [:a {:href "all-commits-list.html"} " See list of all commits"]
        ]