+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in this file. This change
 
 ## [Unreleased]
 ### Added
+- Pie chart for commit statistics (self-committed commits vs. non self-committed commits)
 ### Changed
 ### Removed
 ### Fixed